requirements.txt
absl-py==0.7.1
alabaster==0.7.10
allennlp==0.8.4
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
asn1crypto==0.24.0
astor==0.8.0
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
awscli==1.16.198
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bernoulli==0.1.6
bert-base==0.0.9
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
blis==0.2.4
bokeh==0.12.13
boto==2.48.0
boto3==1.9.43
botocore==1.12.188
Bottleneck==1.2.1
bz2file==0.98
cachetools==2.1.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
conda==4.4.10
conda-build==3.4.1
conda-verify==2.0.0
conllu==0.11
contextlib2==0.5.5
cryptography==2.1.4
cupy==5.3.0
cycler==0.10.0
cymem==2.0.2
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
distributed==1.20.2
docutils==0.14
editdistance==0.5.3
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
fastrlock==0.4
filelock==2.0.13
fire==0.1.3
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
future==0.16.0
gast==0.2.2
gensim==3.6.0
gevent==1.4.0
glob2==0.6
glove-python==0.1.0
gmpy2==2.0.8
google-api-python-client==1.7.4
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-pasta==0.1.7
GPUtil==1.4.0
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.22.0
gunicorn==19.9.0
gym==0.10.5
h5py==2.9.0
heapdict==1.0.0
html5lib==0.9999999
httplib2==0.11.3
idna==2.6
image-classifiers==0.2.0
imageio==2.2.0
imagesize==0.7.1
ipdb==0.12
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.2
jsonnet==0.13.0
jsonpickle==1.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.5
jupyterlab-launcher==0.10.2
Keras==2.2.4
Keras-Applications==1.0.8
keras-bert==0.39.0
keras-bi-lm==0.23.0
keras-contrib==2.0.8
keras-embed-sim==0.3.0
keras-layer-normalization==0.11.0
keras-multi-head==0.18.0
keras-ordered-neurons==0.8.0
keras-pos-embd==0.9.0
keras-position-wise-feed-forward==0.4.0
Keras-Preprocessing==1.1.0
keras-rectified-adam==0.2.0
keras-self-attention==0.39.0
keras-transformer==0.21.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lightgbm==2.1.2
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.3
mock==3.0.5
mpmath==1.0.0
msgpack-python==0.5.1
multipledispatch==0.4.9
murmurhash==1.0.2
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpy==1.16.4
numpydoc==0.9.1
oauth2client==4.1.2
odo==0.5.1
olefile==0.45.1
opencv-python==4.1.0.25
openpyxl==2.4.10
overrides==1.9
packaging==16.8
pandas==0.23.1
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.1
plac==0.9.6
pluggy==0.6.0
ply==3.10
preshed==2.0.1
prettytable==0.7.2
prompt-toolkit==1.0.15
protobuf==3.9.0
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
py4j==0.10.7
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pydevd-pycharm==191.7479.30
pydot==1.4.1
pyflakes==1.6.0
pyglet==1.3.2
Pygments==2.2.0
pylint==1.8.2
pynvrtc==8.0
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyrouge==0.1.3
PySocks==1.6.7
pyspark==2.4.0
pytest==3.3.2
python-dateutil==2.6.1
python-Levenshtein==0.11.2
pytorch-crf==0.7.2
pytorch-pretrained-bert==0.6.1
PyTorch-QRNN==0.2.1
pytorch-transformers==1.0.0
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
regex==2019.3.12
requests==2.18.4
responses==0.10.6
rope==0.10.7
rsa==3.4.2
ruamel-yaml==0.15.35
s3transfer==0.1.13
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
segmentation-models==0.2.1
Send2Trash==1.4.2
sentencepiece==0.1.82
# Editable install with no version control (seq2seq==0.1)
-e /home/ccit/wrp
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.7.1
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
spacy==2.1.7
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
sqlparse==0.3.0
srsly==0.0.7
statsmodels==0.8.0
sympy==1.1.1
synonyms==3.10.2
tables==3.4.2
tblib==1.3.2
tensor2tensor==1.6.6
tensorboard==1.13.1
tensorboardX==1.8
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-hub==0.4.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tf-metrics==0.0.1
Theano==1.0.4
thinc==7.0.8
toolz==0.9.0
torch==1.1.0
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.2.2.post3
tornado==4.5.3
tqdm==4.23.4
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.22
visdom==0.1.8.8
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.5
widgetsnbextension==3.1.0
word2number==1.1
wrapt==1.11.1
xgboost==0.71
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zict==0.1.3
